pyproject.toml
[project]
name = "apollo-io-mcp"
version = "0.1.1"
description = "AThis is an MCP (Model Context Protocol) server specifically built to collect and serve ratings and reviews for other MCP servers deployed on the NetMind platform."
authors = [
    {name = "yanrujing", email = "rujing.yan@protagonist-ai.com"}
]
repository = "https://github.com/protagolabs/Netmind-review-MCP"
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastmcp (>=2.11.3,<3.0.0)"
]

packages = [{ include = "src/apollo_io_mcp" }]


[tool.poetry.scripts]
apollo-io-mcp = "apollo_io_mcp:main"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

